+++ src/applications/pensions/tests/unit/chapters/03-health-and-employment-information/specialMonthlyPension.unit.spec.jsx
@@ -1,16 +1,11 @@
-import { expect } from 'chai';
-import { $ } from '@department-of-veterans-affairs/platform-forms-system/ui';
-
 import {
   testComponentFieldsMarkedAsRequired,
   testNumberOfWebComponentFields,
   testSubmitsWithoutErrors,
-  testShowAlert,
   testNumberOfFieldsByType,
 } from '../pageTests.spec';
 import formConfig from '../../../../config/form';
 import specialMonthlyPension from '../../../../config/chapters/03-health-and-employment-information/specialMonthlyPension';
-import { fillRadio } from '../../testHelpers/webComponents';
 
 const { schema, uiSchema } = specialMonthlyPension;
 
@@ -34,19 +29,6 @@ describe('pension special monthly pension page', () => {
   );
 
   testSubmitsWithoutErrors(formConfig, schema, uiSchema, pageTitle);
-
-  testShowAlert(
-    formConfig,
-    schema,
-    uiSchema,
-    pageTitle,
-    { specialMonthlyPension: false },
-    async container => {
-      const radio = $('va-radio', container);
-      expect(radio).to.exist;
-      await fillRadio(radio, 'Y');
-    },
-  );
 
   testNumberOfFieldsByType(
     formConfig,
