+++ src/applications/pensions/tests/unit/chapters/05-financial-information/transferredAssets.unit.spec.jsx
@@ -1,15 +1,11 @@
-import { $ } from '@department-of-veterans-affairs/platform-forms-system/ui';
-import { expect } from 'chai';
 import {
   testComponentFieldsMarkedAsRequired,
   testNumberOfFieldsByType,
   testNumberOfWebComponentFields,
-  testShowAlert,
   testSubmitsWithoutErrors,
 } from '../pageTests.spec';
 import formConfig from '../../../../config/form';
 import transferredAssets from '../../../../config/chapters/05-financial-information/transferredAssets';
-import { fillRadio } from '../../testHelpers/webComponents';
 
 const { schema, uiSchema } = transferredAssets;
 
@@ -44,18 +40,5 @@ describe('Pensions: Financial information transferred assets page', () => {
       'va-radio': 1,
     },
     pageTitle,
-  );
-
-  testShowAlert(
-    formConfig,
-    schema,
-    uiSchema,
-    pageTitle,
-    { totalNetWorth: false },
-    async container => {
-      const radio = $('va-radio', container);
-      expect(radio).to.exist;
-      await fillRadio(radio, 'Y');
-    },
   );
 });
